# slower calculate_u but less allocs