--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@
 /Manifest.toml
 /docs/build/
 /.CondaPkg/
+.DS_Store