--- a/src/Amica.jl
+++ b/src/Amica.jl
@@ -3,11 +3,12 @@ module Amica
     using LinearAlgebra
 	using GaussianMixtures
 	using Distributions
+    using IntelVectorMath
 	using SpecialFunctions
     using ProgressMeter
     using LoopVectorization
     using AppleAccelerate
-    
+
     include("types.jl")
     include("helper.jl")
     include("likelihood.jl")