… Project.toml
… src/newton.jl
… src/types.jl
… test/bene_testAmica.jl
… test/compare_amica_implementations.jl
… test/profile.jl
… test/runtests.jl